environments/local/services/checksummer/pyproject.toml
[project]
name = "checksummer"
version = "0.1.0"
description = ""
authors = [{ name = "David Sillman", email = "dsillman2000@gmail.com" }]
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "kafka-python (>=2.1.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "crc (>=7.1.0,<8.0.0)",
]

[tool.poetry]
packages = [
    { include = "src" },
    { include = "common", from = "..", to = "common" },
]

[tool.poetry.scripts]
main = "src.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
